api/requirements.txt
flask
pymongo
flask-cors
pyjwt